Update to windowing testing and function
